+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.8.2](https://github.com/openfga/go-sdk/compare/v0.8.1...v0.8.2) (2026-06-08)
+
+
+### Fixed
+
+* process credentials when custom HTTPClient is provided ([#276](https://github.com/openfga/go-sdk/issues/276)) ([07b7c82](https://github.com/openfga/go-sdk/commit/07b7c82679c112c4d9cf9950792b4b7a209af47b))
+
+
+## [0.8.1](https://github.com/openfga/go-sdk/compare/v0.8.0...v0.8.1) (2026-05-20)
+
+### Fixed
+
+* do not retry non-retryable HTTP 4xx client errors ([#319](https://github.com/openfga/go-sdk/issues/319)) ([3d9ff7a](https://github.com/openfga/go-sdk/commit/3d9ff7aeec0941c66d7c271c0b46f28292c3f657))
+
+
 ## [0.8.0](https://github.com/openfga/go-sdk/compare/v0.7.5...v0.8.0) (2026-04-08)
 
 ### Added